Fixed more errors.

+++ module/vision/LUTClassifier/src/QuexClassifier.cpp
@@ -51,8 +51,8 @@ namespace module {
                 size_t length = end[1] - start[1] + 1;
 
                 for(uint i = 0; i < length / subsample; ++i) {
-                    buffer[i + 1] = static_cast<uint8_t>(utility::vision::getPixelColour(lut, 
-                        utility::vision::getPixel(start[0], start[1] + (i * subsample), image.dimensions[0], image.dimensions[1], image.data, static_cast<FOURCC>(image.format))));
+                    buffer[i + 1] = utility::vision::getPixelColour(lut, 
+                        utility::vision::getPixel(start[0], start[1] + (i * subsample), image.dimensions[0], image.dimensions[1], image.data, static_cast<FOURCC>(image.format)));
                 }
 
                 lexer.buffer_fill_region_finish(length / subsample);
@@ -64,8 +64,8 @@ namespace module {
                 size_t length = end[0] - start[0] + 1;
 
                 for(uint i = 0; i < length / subsample; ++i) {
-                    buffer[i + 1] = static_cast<uint8_t>(utility::vision::getPixelColour(lut, 
-                        utility::vision::getPixel(start[0] + (i * subsample), start[1], image.dimensions[0], image.dimensions[1], image.data, static_cast<FOURCC>(image.format))));
+                    buffer[i + 1] = utility::vision::getPixelColour(lut, 
+                        utility::vision::getPixel(start[0] + (i * subsample), start[1], image.dimensions[0], image.dimensions[1], image.data, static_cast<FOURCC>(image.format)));
                 }
 
                 lexer.buffer_fill_region_finish(length / subsample);
